requirements.txt
fastapi
uvicorn[standard]
requests
mercadopago
flask
gspread
google-auth-oauthlib
python-multiprecision
wppconnect-python
